+++ packages/core/src/index.ts
@@ -4,7 +4,7 @@ import type { HoprConnectConfig } from '@hoprnet/hopr-connect'
 
 import { PACKET_SIZE, INTERMEDIATE_HOPS, VERSION, FULL_VERSION } from './constants'
 
-import NetworkPeers from './network/network-peers'
+import NetworkPeers, { Entry } from './network/network-peers'
 import Heartbeat from './network/heartbeat'
 import { findPath } from './path'
 
@@ -489,7 +489,7 @@ class Hopr extends EventEmitter {
     for (const destination of closeChannelDestinations) {
       verbose(`closing ${destination}`)
       try {
-        await this.closeChannel(destination.toPeerId())
+        await this.closeChannel(destination.toPeerId(), 'outgoing')
         verbose(`closed channel to ${destination.toString()}`)
         this.emit('hopr:channel:closed', destination.toPeerId())
       } catch (e) {
@@ -552,7 +552,7 @@ class Hopr extends EventEmitter {
    * @param peer peer to query for, default self
    * @param timeout [optional] custom timeout for DHT query
    */
-  public async getAnnouncedAddresses(peer: PeerId = this.getId(), timeout = 5e3): Promise<Multiaddr[]> {
+  public async getAddressesAnnouncedToDHT(peer: PeerId = this.getId(), timeout = 5e3): Promise<Multiaddr[]> {
     if (peer.equals(this.getId())) {
       return this.libp2p.multiaddrs
     }
@@ -695,6 +695,9 @@ class Hopr extends EventEmitter {
     }
   }
 
+  /**
+   * @returns a list connected peerIds
+   */
   public getConnectedPeers(): PeerId[] {
     if (!this.networkPeers) {
       return []
@@ -702,12 +705,33 @@ class Hopr extends EventEmitter {
     return this.networkPeers.all()
   }
 
+  /**
+   * Takes a look into the indexer.
+   * @returns a list of announced multi addresses
+   */
+  public async getAddressesAnnouncedOnChain(): Promise<Multiaddr[]> {
+    return this.indexer.getAddressesAnnouncedOnChain()
+  }
+
+  /**
+   * @param peerId of the node we want to get the connection info for
+   * @returns various information about the connection
+   */
+  public getConnectionInfo(peerId: PeerId): Entry {
+    return this.networkPeers.getConnectionInfo(peerId)
+  }
+
+  /**
+   * @deprecated Used by API v1
+   * @returns a string describing the connection status between
+   * us and various nodes
+   */
   public async connectionReport(): Promise<string> {
     if (!this.networkPeers) {
       return 'Node has not started yet'
     }
     const connected = this.networkPeers.debugLog()
-    const announced = await this.connector.indexer.getAnnouncedAddresses()
+    const announced = await this.connector.indexer.getAddressesAnnouncedOnChain()
     return `${connected}
     \n${announced.length} peers have announced themselves on chain:
     \n${announced.map((ma: Multiaddr) => ma.toString()).join('\n')}`
@@ -758,7 +782,7 @@ class Hopr extends EventEmitter {
     let addrToAnnounce: Multiaddr
 
     if (announceRoutableAddress) {
-      let multiaddrs = await this.getAnnouncedAddresses()
+      let multiaddrs = await this.getAddressesAnnouncedToDHT()
 
       if (this.options.testing?.announceLocalAddresses) {
         multiaddrs = multiaddrs.filter((ma) => isMultiaddrLocal(ma))
@@ -917,9 +941,15 @@ class Hopr extends EventEmitter {
     }
   }
 
-  public async closeChannel(counterparty: PeerId): Promise<{ receipt: string; status: ChannelStatus }> {
+  public async closeChannel(
+    counterparty: PeerId,
+    direction: 'incoming' | 'outgoing'
+  ): Promise<{ receipt: string; status: ChannelStatus }> {
     const counterpartyPubKey = PublicKey.fromPeerId(counterparty)
-    const channel = await this.db.getChannelX(this.pubKey, counterpartyPubKey)
+    const channel =
+      direction === 'outgoing'
+        ? await this.db.getChannelX(this.pubKey, counterpartyPubKey)
+        : await this.db.getChannelX(counterpartyPubKey, this.pubKey)
 
     // TODO: should we wait for confirmation?
     if (channel.status === ChannelStatus.Closed) {
@@ -968,7 +998,7 @@ class Hopr extends EventEmitter {
   public async getTickets(peerId: PeerId): Promise<Ticket[]> {
     const selfPubKey = new PublicKey(this.getId().pubKey.marshal())
     const counterpartyPubKey = new PublicKey(peerId.pubKey.marshal())
-    const channel = await this.db.getChannelX(selfPubKey, counterpartyPubKey)
+    const channel = await this.db.getChannelX(counterpartyPubKey, selfPubKey)
     return this.db
       .getAcknowledgedTickets({
         channel
@@ -1004,8 +1034,18 @@ class Hopr extends EventEmitter {
   public async redeemTicketsInChannel(peerId: PeerId) {
     const selfPubKey = new PublicKey(this.getId().pubKey.marshal())
     const counterpartyPubKey = new PublicKey(peerId.pubKey.marshal())
-    const channel = await this.db.getChannelX(selfPubKey, counterpartyPubKey)
+    const channel = await this.db.getChannelX(counterpartyPubKey, selfPubKey)
     await this.connector.redeemTicketsInChannel(channel)
+  }
+
+  /**
+   * Get the channel entry between source and destination node.
+   * @param src PeerId
+   * @param dest PeerId
+   * @returns the channel entry of those two nodes
+   */
+  public async getChannel(src: PeerId, dest: PeerId): Promise<ChannelEntry> {
+    return await this.db.getChannelX(new PublicKey(src.pubKey.marshal()), new PublicKey(dest.pubKey.marshal()))
   }
 
   public async getChannelsFrom(addr: Address): Promise<ChannelEntry[]> {